made the adjustments
